+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-### 0.0.2 - in progress
+### 0.0.2 - 2021.07.14
 - queue added
 - duckduckgo search added
+- added microservice to get body from duckduckgo search response and send to queue
 
 ### 0.0.1 - 2021.07.10